Access to the Fipe Database

Access to the Fipe Database

fipe_vehicle(model, make = NULL, year = NULL, date = Sys.Date(),
  progress = FALSE, parallel = FALSE)

Arguments

model
A character vector. Vehicle model name.

make
A character vector. If NULL, search all models in all makes, otherwise only those indicated.

year
A numeric vector. Year of manufacture of the vehicle. If 0 returns vehicles 0 km.

date
A date vector. Reference date for the vehicle price.

progress
A logical, if TRUE print a progress bar.

parallel
A logical, if TRUE apply function in parallel.

Value

A data frame including model, make, year, date and price.

Details

The Fipe Database shows average prices of vehicles on the national market, functioning only as a parameter for negotiations and evaluations. The prices effectively used in negotiations vary according to region, vehicle's conservation, color, accessories or any other factor that might influence the conditions of demand and supply for a specific vehicle. The year of the vehicle refers to the year of the model and are not considered vehicles for professional or special use. The values are expressed in R$ (reais) of the month/year of reference.

Examples

fipe_vehicle(model = "etios", make = "toyota", date = Sys.Date(), year = 0)

#> # A tibble: 12 x 5
#>    model                                    make   year  date       price
#>    <chr>                                    <chr>  <fct> <date>     <dbl>
#>  1 ETIOS PLATINUM 1.5 Flex 16V 4p Aut.      Toyota 0 km  2019-08-01 63127
#>  2 ETIOS PLATINUM Sed. 1.5 Flex 16V 4p Aut. Toyota 0 km  2019-08-01 66080
#>  3 ETIOS X 1.3 Flex 16V 5p Aut.             Toyota 0 km  2019-08-01 54965
#>  4 ETIOS X 1.3 Flex 16V 5p Mec.             Toyota 0 km  2019-08-01 50072
#>  5 ETIOS X Plus 1.5 Flex 16V 5p Aut.        Toyota 0 km  2019-08-01 61448
#>  6 ETIOS X Plus 1.5 Flex 16V 5p Mec.        Toyota 0 km  2019-08-01 56384
#>  7 ETIOS X Plus Sedan 1.5 Flex 16V 4p Aut.  Toyota 0 km  2019-08-01 64180
#>  8 ETIOS X Plus Sedan 1.5 Flex 16V 4p Mec.  Toyota 0 km  2019-08-01 59124
#>  9 ETIOS X Sedan 1.5 Flex 16V 4p Aut.       Toyota 0 km  2019-08-01 59442
#> 10 ETIOS X Sedan 1.5 Flex 16V 4p Mec.       Toyota 0 km  2019-08-01 54264
#> 11 ETIOS XLS 1.5 Flex 5p Aut.               Toyota 0 km  2019-08-01 59526
#> 12 ETIOS XLS Sedan 1.5 Flex 16V 4p Aut.     Toyota 0 km  2019-08-01 59851